Initial hack of main application screens
Continue the work promised in a previous commit.

* `tipjar.xml` is the default page,
* `adjust.xml` is the split info edit page,
* `bill.xml` is the payment page,

also, two views/widgets

* `split_info.xml` will show a person/split on the default page, and be
  used in its ListAdapter
* `person_memo.xml` will show a person and a letter, and be used in the
  GridView on the payment page.

As mentioned in a comment, there's a lot of tedious repetition of XML which
could be refactored into more views, but this is a useful (and more or less
complete) start.  Is that the sound of a waterfall I hear?