Further adjustments to reduce unnecessary recalculation
The model being wired to the edittext like this causes a circular update
problem; set the bill to $10, and the code notices that the difference
between the value in the box (eg, $0.82) and the model ($0.825) is more
than the minimum threshold to update, and so sets it.  Setting the tax
amount defines the tax rate, so this led to the rate losing precision as
you type.  In lieu of improvements to the model (which should be returning
even numbers of cents), avoid this for now by skipping updates if the
difference could be a rounding error.  This will be buggy and result in
cent-level changes to fields sometimes not updating the model.  Also, add a
flag to update which indicates whether a numeric box is possibly being
edited, so that it can update all fields to make them tidy, or only if they
are numerically different, to avoid interfering with input.